Update README.md
This is a resubmission of the ICP Individual Project. The files included in the repository are the class files for the respective files (Airports, Routes and Airlines), as well as class files for Node and one for Main. Every executable code is within the Main class, and I tested my search algorithm using "Accra,Ghana" as the start city and country and "Winnipeg-Canada" as the destination city and country. The output solution path can be found in the winnipeg-canada_output.txt file.